added different soot and new styling on different screen sizes
