+++ src/pages/r101.module.scss
@@ -20,7 +20,7 @@ $dark-grey: #343332;
   }
 }
 
-@media screen and (max-width: 820px) {
+@media screen and (max-width: 855px) {
   .mainStepContainer {
     flex-wrap: wrap;
   }
